Traject::ThreadPool basedon Concurrent::ThreadPool
is the only Traject::ThreadPool we need, no need for separate
jruby implementation, Concurrent::ThreadPool already takes care
of that for us.
